Split efinder.py in 2, capture with fake camera works
